services/runner/pyproject.toml
[project]
name = "talaty-runner"
version = "0.1.0"
description = "Execution runner for the Talaty scoring platform."
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6,<3",
  "pydantic-settings>=2.2,<3",
  "pyyaml>=6.0,<7",
  "sqlalchemy>=2.0,<3",
  "psycopg[binary]>=3.1,<4",
  "python-dotenv>=1.0,<2",
  "boto3>=1.34,<2",
  "fastapi>=0.111,<1",
  "uvicorn[standard]>=0.30,<1",
  "mlflow>=2.11,<3",
]

[project.scripts]
talaty-runner = "runner.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

